requirements.txt
openai==1.20.0
python-dotenv
yfinance
pillow
tk